Create max_favorite_songs
Here is the text from the image:

---

Alice has a collection of songs represented as a string S where each character represents a song. A playlist is the substring of the given string with exactly k number of songs. She wants to create a playlist that contains the maximum number of her favourite song which is 'a'. Your task is to find and return an integer value representing the maximum number of favourite songs that she can get in a single playlist.

Input:
Abcaca  
3

Output:
2